Update README.md
In our daily life, people commonly face many problems in finding which path leads. People usually explore every possible solution in finding their way within the campus, but not every solution can produce the optimal path. Shortest path problem is a problem in finding the fastest route or path from a directed graph. The “Campus Navigation” is used for navigating the user from one place to another within the campus. All navigation system takes user location as a starting point and gives the best optimal path to be followed. Every direction in a graph (map) has a cost to be calculated. This shortest path problem is a way to find a new route or path in a graph with a minimum sum of weight travelled through the direction. This shortest path problem is solved by using Dijkstra’s algorithm of finding the best edge path between vertices in a graph. There are several variations of algorithm that can be used to determine the node that was pursued based on the direction in given graph. Variations of the shortest path can be distinguished from single-source objective, pair path and generalization. A pair of shortest paths is finding the shortest path for two points of nodes. All pair of shortest paths is a technique to find the shortest path among all directed nodes. Single-source shortest path is finding the shortest form travelled, starting from a certain node to all other nodes in the graph. Single-objective shortest path problem is to find the shortest path from any node on the graph are directed to a single destination node. Intermediate shortest path is finding the shortest path between two nodes selected through other nodes. Generalization is significantly more efficient than the simple approach to run one-pair of shortest path algorithm on all pairs of vertices that are relevant.